Move template components to pkg/lib/web
